requirements.txt
tqdm
numpy==1.19.5
pandas
pyyaml==5.4
Sphinx==3.3.1
sphinxcontrib-serializinghtml
rinohtype
py-cpuinfo
pytest
graphviz
sklearn
colorama
pyperf